feat: Reintroduce judge worker functionality with dispute processing
